Initial Setup for Displaycommunication
Display can be devided in zones. Every Zone has it own properties:
Position
width
height
textColor
backgroundColor
content or value

Content and Value can be changed after creation via a function. With the Adafruit_SSD1306 Library and compatible display, the Zones can be printet/shown directly.
The rectangle of a zone can be inverted with an function.